+++ web/src/io/getContributors.ts
@@ -1,16 +1,5 @@
-/* eslint-disable camelcase */
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import { contributors } from 'src/../../.all-contributorsrc'
-
-export interface ContributorData {
-  login: string
-  name: string
-  avatar_url: string
-  profile: string
-  contributions: string[]
-}
+import allContributors, { ContributorData } from 'src/../../.all-contributorsrc'
 
 export function getContributors(): ContributorData[] {
-  return contributors
+  return allContributors.contributors
 }
