+++ web/src/types/all-contributorsrc.d.ts
@@ -1,0 +1,13 @@
+/* eslint-disable camelcase */
+declare module '*.all-contributorsrc' {
+  export interface ContributorData {
+    login: string
+    name: string
+    avatar_url: string
+    profile: string
+    contributions: string[]
+  }
+
+  const json: { contributors: ContributorData[] }
+  export default json
+}
